requirements.txt
asgiref==3.7.2
attrs==23.1.0
build==0.10.0
CacheControl==0.13.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cleo==2.0.1
crashtest==0.4.1
cryptography==41.0.2
distlib==0.3.7
Django==4.2.5
django-background-task==0.1.8
djangorestframework==3.14.0
dulwich==0.21.5
filelock==3.12.2
idna==3.4
importlib-metadata==6.7.0
installer==0.7.0
jaraco.classes==3.2.3
jeepney==0.8.0
jsonschema==4.17.3
keyring==24.2.0
more-itertools==9.1.0
msgpack==1.0.5
packaging==23.1
pexpect==4.8.0
pipenv==2023.8.26
pkginfo==1.9.6
platformdirs==3.10.0
poetry==1.6.1
poetry-core==1.7.0
poetry-plugin-export==1.5.0
ptyprocess==0.7.0
pycparser==2.21
pyproject_hooks==1.0.0
pyrsistent==0.19.3
pytz==2023.3.post1
rapidfuzz==2.15.1
requests==2.31.0
requests-toolbelt==1.0.0
SecretStorage==3.3.3
shellingham==1.5.0.post1
sqlparse==0.4.4
tomli==2.0.1
tomlkit==0.11.8
trove-classifiers==2023.5.24
typing_extensions==4.8.0
urllib3==1.26.16
virtualenv==20.24.3
zipp==3.15.0
